fix: busy loop on shutdown (#30)
* tests: simplify window_size_decremented_past_zero (#829)

This helps finding the root cause of why the assertion
in FlowControl::dec_send_window failed in the first place.

See #692 for more details.

* Fix reclaiming reserved capacity (#832)

Reclaiming requested capacity that has not been actually reserved yet
is wrong, as this capacity never existed to begin with.

Fixes #607

* fix: busy loop on shutdown (#834)

When in a slow shutdown, this was busy looping since no more streams were left. The change here now only makes sure we poll one extra time if the state changed during the poll.

Fixes #831

---------

Co-authored-by: Anthony Ramine <123095+nox@users.noreply.github.com>
Co-authored-by: Sean McArthur <sean@seanmonstar.com>